+++ src/main/decorators/log.spec.ts
@@ -1,6 +1,6 @@
 import { Controller, HttpRequest, httpResponse } from '@presentation/protocols'
 import { LogControllerDecorator } from './log'
-import { serverError, ok } from '@presentation/helpers/http-helper'
+import { serverError, ok } from '@presentation/helpers/http/http-helper'
 import { LogErrorRepository } from '@data/protocols/log-error-repository'
 import { AccountModel } from '@domain/models/account'
