+++ src/presentation/controller/login/login.spec.ts
@@ -1,5 +1,5 @@
 import { MissingParamError } from '@presentation/errors'
-import { badRequest, ok, serverError, unauthorized } from '@presentation/helpers/http-helper'
+import { badRequest, ok, serverError, unauthorized } from '@presentation/helpers/http/http-helper'
 import { LoginController } from './login'
 import { Authentication, HttpRequest, Validation } from './login-protocols'
